+++ controllers/contacts.js/getById.js
@@ -1,0 +1,18 @@
+const contactsOperation = require('../../model');
+
+const getById = async (req, res, next) => {
+  try {
+    const { contactId } = req.params;
+    const contact = await contactsOperation.getContactById(contactId);
+    if (!contact) {
+      const error = new Error('Not found');
+      error.status = 404;
+      throw error;
+    }
+    res.json(contact);
+  } catch (error) {
+    next(error);
+  }
+};
+
+module.exports = getById;
